+++ sources/Runner/LocalIncludes.hpp
@@ -1,3 +1,4 @@
 #include "Config.hpp"
+#include "ConfigJson.hpp"
 #include "LibmodbusAbstraction.hpp"
 #include "ModbusTechnologyAdapter.hpp"