+++ test_package/LocalIncludes.hpp
@@ -1,3 +1,4 @@
 #include "Modbus_Technology_Adapter/Config.hpp"
+#include "Modbus_Technology_Adapter/ConfigJson.hpp"
 #include "Modbus_Technology_Adapter/LibmodbusAbstraction.hpp"
 #include "Modbus_Technology_Adapter/ModbusTechnologyAdapter.hpp"