+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
+### Added
+- module `ConfigJson`
+
 ## [0.0.0] - 2023.05.08